+++ src/collections/Bookings/index.ts
@@ -5,9 +5,6 @@ import { slugField } from '@/fields/slug'
 import type { CollectionConfig } from 'payload'
 import { hostOrSelfOrGuests } from './access/adminOrSelfOrGuests'
 import { generateJwtToken, verifyJwtToken } from '@/utilities/token'
-import { unavailableDates } from './endpoints/unavailable-dates'
-import { checkAvailability } from './endpoints/check-availability'
-import { checkAvailabilityHook } from './hooks/checkAvailability'
 
 export const Booking: CollectionConfig = {
   slug: 'bookings',
@@ -30,8 +27,6 @@ export const Booking: CollectionConfig = {
     useAsTitle: 'customer',
   },
   endpoints: [
-    unavailableDates,
-    checkAvailability,
     // This endpoint is used to generate a token for the booking
     // and return it to the customer
     {
@@ -458,30 +453,45 @@ export const Booking: CollectionConfig = {
     create: ({ req: { user } }) => {
       if (!user) return false
       const roles = user.role || []
-      return roles.includes('admin') || roles.includes('customer')
+      // Only hosts and customers can create bookings, guests cannot
+      return roles.includes('host') || roles.includes('customer')
     },
     read: ({ req: { user } }) => {
       if (!user) return false
-      if (user.role?.includes('admin')) return true
+      if (user.role?.includes('host')) return true
+      
+      // Build conditions for customers and guests
+      const conditions: any[] = []
+      
       if (user.role?.includes('customer')) {
-        return { customer: { equals: user.id } }
+        conditions.push({ customer: { equals: user.id } })
       }
-      return false
+      
+      if (user.role?.includes('guest')) {
+        conditions.push({ guests: { contains: user.id } })
+      }
+      
+      if (conditions.length === 0) return false
+      if (conditions.length === 1) return conditions[0]
+      
+      return { or: conditions }
     },
     update: ({ req: { user } }) => {
       if (!user) return false
-      if (user.role?.includes('admin')) return true
+      if (user.role?.includes('host')) return true
       if (user.role?.includes('customer')) {
         return { customer: { equals: user.id } }
       }
+      // Guests cannot update bookings
       return false
     },
     delete: ({ req: { user } }) => {
       if (!user) return false
-      if (user.role?.includes('admin')) return true
+      if (user.role?.includes('host')) return true
       if (user.role?.includes('customer')) {
         return { customer: { equals: user.id } }
       }
+      // Guests cannot delete bookings
       return false
     },
   },
@@ -499,11 +509,11 @@ export const Booking: CollectionConfig = {
       name: 'customer',
       type: 'relationship',
       relationTo: 'users',
-      // filterOptions: {
-      //   role: {
-      //     equals: 'customer',
-      //   },
-      // },
+      filterOptions: {
+        role: {
+          equals: 'customer',
+        },
+      },
       access: {
         update: isHostField,
       },
@@ -512,7 +522,7 @@ export const Booking: CollectionConfig = {
       name: 'token',
       label: 'Token',
       type: 'text',
-      required: false,
+      required: true,
       admin: {
         readOnly: true,
         hidden: true,
